pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "getemails"
version = "0.1.0"
description = "Download and archive emails from Gmail, iCloud, and AOL to .eml files"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.0",
    "google-auth-httplib2>=0.2",
    "google-auth-oauthlib>=1.0",
    "imapclient>=2.3",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "click>=8.1",
]

[project.scripts]
getemails = "getemails.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["getemails"]